+++ temba/contacts/models.py
@@ -1790,47 +1790,38 @@ class Contact(RequireUpdateFieldsMixin, TembaModel):
             user = get_anonymous_user()
         self.unstop(user)
 
-    def deactivate(self, user):
-        if self.is_active:
-            with transaction.atomic():
-
-                # prep our urns for deletion so our old path creates a new urn
-                for urn in self.urns.all():
-                    path = str(uuid.uuid4())
-                    urn.identity = f"{DELETED_SCHEME}:{path}"
-                    urn.path = path
-                    urn.scheme = DELETED_SCHEME
-                    urn.channel = None
-                    urn.save(update_fields=("identity", "path", "scheme", "channel"))
-
-                # no group for you!
-                self.clear_all_groups(user)
-
-                # now deactivate the contact itself
-                self.is_active = False
-                self.name = None
-                self.fields = None
-                self.modified_by = user
-                self.save(update_fields=("name", "is_active", "fields", "modified_on", "modified_by"))
-
-    def release_async(self, user):
+    def release(self, user, *, immediately=True):
         """
         Marks this contact for deletion
         """
+        with transaction.atomic():
+            # prep our urns for deletion so our old path creates a new urn
+            for urn in self.urns.all():
+                path = str(uuid.uuid4())
+                urn.identity = f"{DELETED_SCHEME}:{path}"
+                urn.path = path
+                urn.scheme = DELETED_SCHEME
+                urn.channel = None
+                urn.save(update_fields=("identity", "path", "scheme", "channel"))
 
-        # mark us as inactive
-        self.deactivate(user)
+            # no group for you!
+            self.clear_all_groups(user)
+
+            # now deactivate the contact itself
+            self.is_active = False
+            self.name = None
+            self.fields = None
+            self.modified_by = user
+            self.save(update_fields=("name", "is_active", "fields", "modified_on", "modified_by"))
 
         # kick off a task to remove all the things related to us
-        from temba.contacts.tasks import release_contact
+        if immediately:
+            from temba.contacts.tasks import full_release_contact
 
-        release_contact.delay(self.id, user.id)
+            full_release_contact.delay(self.id, user.id)
 
-    def release(self, user):
+    def _full_release(self, user):
         with transaction.atomic():
-
-            # make sure we've been deactivated
-            self.deactivate(user)
 
             # release our messages
             for msg in self.msgs.all():
@@ -2056,6 +2047,9 @@ class Contact(RequireUpdateFieldsMixin, TembaModel):
         # perform everything in a org-level lock to prevent duplication by different instances. Org-level is required
         # to prevent conflicts with get_or_create which uses an org-level lock.
 
+        # list of other contacts that were modified
+        modified_contacts = set()
+
         with self.org.lock_on(OrgLock.contacts):
 
             # urns are submitted in order of priority
@@ -2069,8 +2063,11 @@ class Contact(RequireUpdateFieldsMixin, TembaModel):
                     urn = ContactURN.create(self.org, self, normalized, priority=priority)
                     urns_created.append(urn)
 
-                # unassigned URN or assigned to someone else
+                # unassigned URN or different contact
                 elif not urn.contact or urn.contact != self:
+                    if urn.contact:
+                        modified_contacts.add(urn.contact.id)
+
                     urn.contact = self
                     urn.priority = priority
                     urn.save()
@@ -2096,6 +2093,10 @@ class Contact(RequireUpdateFieldsMixin, TembaModel):
 
         # trigger updates based all urns created or detached
         self.handle_update(urns=[str(u) for u in (urns_created + urns_attached + urns_detached)])
+
+        # update modified on any other modified contacts
+        if modified_contacts:
+            Contact.objects.filter(id__in=modified_contacts).update(modified_on=timezone.now())
 
         # clear URN cache
         if hasattr(self, "__urns"):
@@ -2879,6 +2880,7 @@ class ContactGroup(TembaModel):
         self.is_active = False
         self.save()
         self.contacts.clear()
+        self.counts.all().delete()
 
         # delete any event fires related to our group
         from temba.campaigns.models import EventFire